+++ src/main/cpp/util/field_utilities.h
@@ -118,7 +118,7 @@ namespace vaultdb {
         }
 
         static inline emp::Integer padInteger(const emp::Integer & src, const int & len) {
-            if(src.size() == len) return src;
+            if(src.size() == (size_t) len) return src;
 
             emp::Integer dst(len, 0, PUBLIC);
             memcpy(dst.bits.data(), src.bits.data(), src.size() * TypeUtilities::getEmpBitSize());
